requirements.txt
contourpy==1.3.2
cycler==0.12.1
diff_jpeg @ git+https://github.com/necla-ml/Diff-JPEG@e81f082896ba145e35cc129bc7743244e10881e5
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
Jinja2==3.1.6
kiwisolver==1.4.9
lpips==0.1.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyYAML==6.0.3
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing_extensions==4.15.0
